backend/requirements.txt
# Testing and data source dependencies for CI
pytest==8.0.0
sec-edgar-downloader==5.2.0
fastapi==0.109.1
uvicorn[standard]==0.27.0
neo4j==5.16.0
networkx==3.2.1
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.27.0
edgartools==4.4.0  # For SEC EDGAR filings and extractions
usaspending-orm==0.2.0  # For USA Spending contract queries